+++ src/main/java/com/foodielog/server/admin/type/BlockReason.java
@@ -1,0 +1,4 @@
+package com.foodielog.server.admin.type;
+
+public enum BlockReason {
+}
