+++ src/main/java/com/foodielog/server/admin/type/ProcessedStatus.java
@@ -1,0 +1,7 @@
+package com.foodielog.server.admin.type;
+
+public enum ProcessedStatus {
+	UNPROCESSED,
+	APPROVED,
+	REJECTED;
+}
