--- src/main/java/com/foodielog/server/application/feed/entity/FeedLike.java
@@ -1,45 +1,0 @@
-package com.foodielog.server.application.feed.entity;
-
-import java.sql.Timestamp;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.CreationTimestamp;
-
-import com.foodielog.server.application.notification.entity.Notification;
-import com.foodielog.server.application.user.entity.User;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "feed_like_tb")
-@Entity
-public class FeedLike {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne
-	@JoinColumn(name = "feed_id")
-	private Feed feed;
-
-	@ManyToOne
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@ManyToOne
-	@JoinColumn(name = "notification")
-	private Notification notification;
-
-	@CreationTimestamp
-	private Timestamp createAt;
-}
