--- src/main/java/com/foodielog/server/application/feed/type/ContentStatus.java
@@ -1,7 +1,0 @@
-package com.foodielog.server.application.feed.type;
-
-public enum ContentStatus {
-	NORMAL,
-	HIDDEN,
-	DELETED;
-}
