--- src/main/java/com/foodielog/server/application/reply/entity/Reply.java
@@ -1,56 +1,0 @@
-package com.foodielog.server.application.reply.entity;
-
-import java.sql.Timestamp;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import com.foodielog.server.application.feed.entity.Feed;
-import com.foodielog.server.application.feed.type.ContentStatus;
-import com.foodielog.server.application.user.entity.User;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "reply_tb")
-@Entity
-public class Reply {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "feed_id")
-	private Feed feed;
-
-	@Column(length = 150, nullable = false)
-	private String content;
-
-	@Enumerated(EnumType.STRING)
-	private ContentStatus status;
-
-	@CreationTimestamp
-	private Timestamp createdAt;
-
-	@UpdateTimestamp
-	private Timestamp updatedAt;
-}
