--- src/main/java/com/foodielog/server/application/report/type/ReportReason.java
@@ -1,9 +1,0 @@
-package com.foodielog.server.application.report.type;
-
-public enum ReportReason {
-	광고,
-	욕설,
-	명예훼손,
-	음란,
-	기타;
-}
