--- src/main/java/com/foodielog/server/application/report/type/ReportType.java
@@ -1,6 +1,0 @@
-package com.foodielog.server.application.report.type;
-
-public enum ReportType {
-	FEED,
-	REPLY;
-}
