--- src/main/java/com/foodielog/server/application/restaurant/entity/RestaurantLike.java
@@ -1,40 +1,0 @@
-package com.foodielog.server.application.restaurant.entity;
-
-import java.sql.Timestamp;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.CreationTimestamp;
-
-import com.foodielog.server.application.user.entity.User;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "restaurant_like_tb")
-@Entity
-public class RestaurantLike {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne
-	@JoinColumn(name = "restaurant_id")
-	private Restaurant restaurant;
-
-	@ManyToOne
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@CreationTimestamp
-	private Timestamp createdAt;
-}
