--- src/main/java/com/foodielog/server/application/user/type/Flag.java
@@ -1,6 +1,0 @@
-package com.foodielog.server.application.user.type;
-
-public enum Flag {
-	YES,
-	NO;
-}
