--- src/main/java/com/foodielog/server/application/user/type/Role.java
@@ -1,6 +1,0 @@
-package com.foodielog.server.application.user.type;
-
-public enum Role {
-	USER,
-	ADMIN;
-}
