--- src/main/java/com/foodielog/server/application/user/type/UserStatus.java
@@ -1,7 +1,0 @@
-package com.foodielog.server.application.user.type;
-
-public enum UserStatus {
-	NORMAL,
-	WITHDRAW,
-	BLOCK;
-}
