+++ src/main/java/com/foodielog/server/feed/type/ContentStatus.java
@@ -1,0 +1,7 @@
+package com.foodielog.server.feed.type;
+
+public enum ContentStatus {
+	NORMAL,
+	HIDDEN,
+	DELETED;
+}
