--- src/main/java/com/foodielog/server/management/admin/entity/BadgeApply.java
@@ -1,47 +1,0 @@
-package com.foodielog.server.management.admin.entity;
-
-import java.sql.Timestamp;
-
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import com.foodielog.server.management.admin.type.ProcessedStatus;
-import com.foodielog.server.application.user.entity.User;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "badge_apply_tb")
-@Entity
-public class BadgeApply {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@Enumerated(EnumType.STRING)
-	private ProcessedStatus status;
-
-	@CreationTimestamp
-	private Timestamp createdAt;
-
-	@UpdateTimestamp
-	private Timestamp updatedAt;
-}
