--- src/main/java/com/foodielog/server/management/admin/entity/BlockUser.java
@@ -1,52 +1,0 @@
-package com.foodielog.server.management.admin.entity;
-
-import java.sql.Timestamp;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.DynamicInsert;
-
-import com.foodielog.server.management.admin.type.BlockReason;
-import com.foodielog.server.application.user.entity.User;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@DynamicInsert
-@Table(name = "blockUser_tb")
-@Entity
-public class BlockUser {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@OneToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@Enumerated(EnumType.STRING)
-	private BlockReason reason;
-
-	@Column(name = "feed_count")
-	private Long feedCount;
-
-	@Column(name = "reply_count")
-	private Long replyCount;
-
-	@CreationTimestamp
-	private Timestamp createdAt;
-}
