--- src/main/java/com/foodielog/server/management/admin/entity/WithDrawUser.java
@@ -1,44 +1,0 @@
-package com.foodielog.server.management.admin.entity;
-
-import java.sql.Timestamp;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.CreationTimestamp;
-
-import com.foodielog.server.application.user.entity.User;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "withdraw_user_tb")
-@Entity
-public class WithDrawUser {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@OneToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@Column(name = "feed_count")
-	private Long feedCount;
-
-	@Column(name = "reply_count")
-	private Long replyCount;
-
-	@CreationTimestamp
-	private Timestamp createdAt;
-}
