--- src/main/java/com/foodielog/server/management/admin/type/BlockReason.java
@@ -1,4 +1,0 @@
-package com.foodielog.server.management.admin.type;
-
-public enum BlockReason {
-}
