--- src/main/java/com/foodielog/server/management/admin/type/ProcessedStatus.java
@@ -1,7 +1,0 @@
-package com.foodielog.server.management.admin.type;
-
-public enum ProcessedStatus {
-	UNPROCESSED,
-	APPROVED,
-	REJECTED;
-}
