--- src/main/java/com/foodielog/server/management/admin/type/WithDrawReason.java
@@ -1,5 +1,0 @@
-package com.foodielog.server.management.admin.type;
-
-public enum WithDrawReason {
-
-}
