+++ src/main/java/com/foodielog/server/report/type/ReportReason.java
@@ -1,0 +1,9 @@
+package com.foodielog.server.report.type;
+
+public enum ReportReason {
+	광고,
+	욕설,
+	명예훼손,
+	음란,
+	기타;
+}
