+++ src/main/java/com/foodielog/server/report/type/ReportType.java
@@ -1,0 +1,6 @@
+package com.foodielog.server.report.type;
+
+public enum ReportType {
+	FEED,
+	REPLY;
+}
