+++ src/main/java/com/foodielog/server/user/type/UserStatus.java
@@ -1,0 +1,7 @@
+package com.foodielog.server.user.type;
+
+public enum UserStatus {
+	NORMAL,
+	WITHDRAW,
+	BLOCK;
+}
